fix test

+++ component/broadcast_test.go
@@ -15,7 +15,7 @@ func TestBroadcastComponent_GetBroadcast(t *testing.T) {
 
 	broadcast := database.Broadcast{ID: 1, Content: "test", BcType: "banner", Theme: "light", Status: "active"}
 
-	cc.mocks.stores.BroadcastMock().EXPECT().Get(ctx, 1).Return(
+	cc.mocks.stores.BroadcastMock().EXPECT().Get(ctx, int64(1)).Return(
 		&broadcast, nil,
 	)
 	data, err := cc.GetBroadcast(ctx, 1)
